Class File Refactor, MenuBar Display
Refactored the various components of old mainView.py into different classes.
loginView and mainView are views in program, masterView is view interface.

MenuBar has been implemented in masterView (it has to be added to root Tk frame), can display and hide based on view switched to.
-Cascades implemented = File, Profile, View, Help
(All of these currently print text for validation, some have functionality)
-File ->Exit works
-Profile->LogOut works
-View->any listed view works (cannot test until other views implemented)
-Help->prints text
@TODO login protocol will need to cleanup any stored data unless specified by user
